requirements.txt
paho-mqtt
curses